docker: Rename siikr.conf -> siikr.example.env
Since siikr.conf is more of an environment file, the extension should
reflect that. I have also renamed it to siikr.example.env - this is so
that if the configuation is changed, it will only go into the example
env file, and not cause issues with pulling down into a modified
siikr.env file. Additionally, it helps prevent in-dev configuration from
accidentally getting committed in the case that someone uses
`git commit -a` a bit carelessly.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>