decentralization code along with new features.
new features include:
- more control over advanced search parameters
- better post previews
- feedback on blog existence / proper input
- hopefully more agnostic to postgres permissions

decentralization features include:
- server caching
- remote SSE streaming.
- ingest delegation.
- all spokes are also hubs.

known issues (contributions appreciated):
- a consequence of this quick and dirty approach to decentrailization is that in theory a sufficiently dedicated voyeur can in theory sometimes infer roughly which search must be being carried out on which blogs by snooping in on the search results being broadcast to anyone who wishes to listen for them. Either serverEvents.php needs to be more selective about which clients it forwards messages to, or else result broadcasts should be encrypted by some scheme that plays nicely with no-CORS SSE.
- the server now has sufficient information for the client to show all types of post previews, however, clientside functionality for video, audio, and poll previews has yet to be implemented.

Please bear with all of the commented out debugging code for now. I know it's a bit annoying, but I'll probably need it until the next version.

There is also some code which will become relevant in the next version for decentralized word clouds. I will add a prompt to the setup wizard to allow people to opt out of hosting wordcloud data as it requres overhead (but for consistency either wordcloud participation will be required in order to serve as a spoke, or otherwise I will need to come up with some mechanism that allows a node to host a blog while requesting a seperate node hosts its wordcloud).